% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxgradientboost.R
\name{cox_gradient_boost}
\alias{cox_gradient_boost}
\title{Cox Gradient Boosting Model}
\usage{
cox_gradient_boost(
  X,
  time,
  event,
  regr_fun,
  M = 100L,
  nu = 0.1,
  show_progress = TRUE,
  ...
)
}
\arguments{
\item{X}{numeric matrix of covariates (n x p)}

\item{time}{numeric vector of observed times}

\item{event}{integer/logical vector (1 = event, 0 = censored)}

\item{regr_fun}{function(X, y, ...) -> model with predict() method}

\item{M}{number of boosting iterations (default 100)}

\item{nu}{learning rate / shrinkage (default 0.1)}

\item{show_progress}{logical; print a progress bar? (default TRUE)}

\item{...}{extra arguments forwarded to regr_fun at every iteration}
}
\value{
object of class "CoxGradientBoost"
}
\description{
Fit a model-agnostic Cox-gradient boosting survival model
}
\examples{

require(glmnet)
require(survival)

data(ovarian)

set.seed(42)
idx_train <- sample(nrow(ovarian), floor(0.75 * nrow(ovarian)))
df <- ovarian
train <- df[idx_train, ]; test <- df[-idx_train, ]

regr_lm <- function(X, y, ...) lm(y ~ ., data = data.frame(X, y = y))

fit_boost_lm <- survivalisttoo::cox_gradient_boost(train, train$futime, 
train$fustat, regr_lm)

y_test   <- Surv(test$futime, test$fustat)
(ci_blm <- glmnet::Cindex(predict(fit_boost_lm, test), y_test)) # C-index

}
