% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxgradientboost.R
\name{print.CoxGradientBoost}
\alias{print.CoxGradientBoost}
\title{Print a CoxGradientBoost Object}
\usage{
\method{print}{CoxGradientBoost}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"CoxGradientBoost"}.}

\item{...}{Further arguments passed to or from other methods (currently unused).}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Displays a concise summary of a fitted \code{CoxGradientBoost} model,
including the base learner, number of boosting iterations, learning rate,
and basic information about the training data.
}
\details{
This method prints key characteristics of the fitted model:
\itemize{
  \item The base learner used in boosting
  \item The number of boosting iterations (\eqn{M})
  \item The learning rate (\eqn{\nu})
  \item The number of observations
  \item The number of observed events
  \item The number of unique event times
}
}
\seealso{
\code{\link{cox_gradient_boost}}, \code{\link{predict.CoxGradientBoost}}
}
