\name{survivalisttoo-package}
\alias{survivalisttoo-package}
\alias{survivalisttoo}
\docType{package}
\title{\packageTitle{survivalisttoo}}
\description{\packageDescription{survivalisttoo}}
\section{Package Content}{\packageIndices{survivalisttoo}}
\author{\packageAuthor{survivalisttoo}}
\section{Maintainer}{\packageMaintainer{survivalisttoo}}
\keyword{package}
